[Refactor] 전체 조회 메서드 추가
