+++ src/main/java/com/nhnacademy/bookpubshop/author/repository/AuthorRepository.java
@@ -1,8 +1,6 @@
 package com.nhnacademy.bookpubshop.author.repository;
 
-import com.nhnacademy.bookpubshop.author.dto.GetAuthorResponseDto;
 import com.nhnacademy.bookpubshop.author.entity.Author;
-import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 /**
@@ -11,6 +9,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
  * @author : 여운석
  * @since : 1.0
  **/
-public interface AuthorRepository extends JpaRepository<Author, Integer>, AuthorRepositoryCustom {
-    List<GetAuthorResponseDto> findAuthorsByAuthorName(String name);
+public interface AuthorRepository
+        extends JpaRepository<Author, Integer>, AuthorRepositoryCustom {
+
 }
